move StoppableThread to own file
